Bug 1351480 - Update webrender to 7463ae5908ca1d4065763a8753af5d72a6f78b85. r=jrmuizel
In addition to updating gfx/webrender and gfx/webrender_traits, as well as the
Cargo.toml/Cargo.lock files and third-party rust dependencies, this updates
webrender_bindings for the following API changes:
- Update call to update_image for change in 086f912ea86d053d95e5f4eec00cae0d82ee5dba
- Update request_blob_image for new parameter added in 086f912ea86d053d95e5f4eec00cae0d82ee5dba.
- set_root_display_list got renamed to set_display_list, so update call sites.

+++ gfx/webrender/src/clip_scroll_tree.rs
@@ -23,7 +23,7 @@ pub struct ClipScrollTree {
     /// The current reference frame id, used for giving a unique id to all new
     /// reference frames. The ClipScrollTree increments this by one every time a
     /// reference frame is created.
-    current_reference_frame_id: usize,
+    current_reference_frame_id: u64,
 
     /// The root reference frame, which is the true root of the ClipScrollTree. Initially
     /// this ID is not valid, which is indicated by ```node``` being empty.
